Move tests into parser_impl, remove dead code
